auth finalizado

+++ app/Http/Controllers/IdeaController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\StoreIdeaRequest;
 use App\Models\Idea;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\View\View;
 
 class IdeaController extends Controller
@@ -29,30 +30,37 @@ class IdeaController extends Controller
         $user = Auth::user();
         $user->ideas()->create($request->validated());
 
+        //notify the user that the idea was created successfully
+        
+
         return redirect()->route('ideas.index');
     }
 
     public function show(Idea $idea): View
     {
+        Gate::authorize("manage-idea", $idea);
         return view('ideas.show', ['idea' => $idea]);
     }
 
     public function edit(Idea $idea): View
     {
+        Gate::authorize("manage-idea", $idea);
         return view('ideas.edit', ['idea' => $idea]);
     }
 
     public function update(StoreIdeaRequest $request, Idea $idea): RedirectResponse
     {
+        Gate::authorize("manage-idea", $idea);
         /** @var \App\Models\User $user */
         $user = Auth::user();
-        $user->ideas()->update($request->validated());
+        $user->ideas()->findOrFail($idea->id)->update($request->validated());
 
         return redirect()->route('ideas.show', $idea);
     }
 
     public function destroy(Idea $idea): RedirectResponse
     {
+        Gate::authorize("manage-idea", $idea);
         $idea->delete();
 
         return redirect()->route('ideas.index');
